third_party/blink/web_tests/fast/js/toString-dontEnum-expected.txt

This tests that the toString() function does not enumerate.

PASS: the toString function is not enumerable for HTMLDivElement.
PASS: the toString function is not enumerable for HTMLDocument.
PASS: the toString function is not enumerable for Object.
PASS: the toString function is not enumerable for Array.
